Unset `expires_in` when queueing jobs
The `expires_in` field is meant to be a duration after which the job will expire.
This value gets converted into a timestamp that is stored as `expires_at`.

Having both `expires_in` and `expires_at` set when queueing a job would be ambiguous,
and is therfore an error. Unfortunately we were not following the same rules when
parsing `expires_in` and jobs with that set were being queued with both fields. This
did not play well with the `requeue` plugin which requeues jobs exactly as they're
defined, and would then encounter the error about the ambiguity.

Unsetting `expires_in` after it has been converted to a timestamp should resolve
the issue.